requirements.txt
astroid==2.11.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
Deprecated==1.2.13
dill==0.3.5.1
Flask==2.2.2
gcloud==0.18.3
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.3.1
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
numpy==1.23.1
oauth2client==4.1.3
pandas==1.4.3
platformdirs==2.5.2
plotly==5.9.0
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pylint==2.14.5
pyparsing==3.0.9
Pyrebase4==4.5.0
python-dateutil==2.8.2
python-jwt==3.3.3
pytz==2022.1
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
tenacity==8.0.1
tomli==2.0.1
tomlkit==0.11.3
urllib3==1.26.11
Werkzeug==2.2.2
wrapt==1.14.1
